# shared/openapi/server-api.yaml
openapi: 3.1.0
info:
  title: Power Monitor Server API
  description: Central server for pm-protocol/1.0.0 sensor fleets
  version: 1.0.0
paths:
  /api/v1/auth/bootstrap:
    post:
      tags:
      - authentication
      summary: Bootstrap
      operationId: bootstrap_api_v1_auth_bootstrap_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BootstrapRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionView'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/login:
    post:
      tags:
      - authentication
      summary: Login
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionView'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/session:
    get:
      tags:
      - authentication
      summary: Session Status
      operationId: session_status_api_v1_auth_session_get
      parameters:
      - name: pm_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pm Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionView'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/logout:
    post:
      tags:
      - authentication
      summary: Logout
      operationId: logout_api_v1_auth_logout_post
      parameters:
      - name: X-CSRF-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Csrf-Token
      - name: pm_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pm Session
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/reauthenticate:
    post:
      tags:
      - authentication
      summary: Reauthenticate
      operationId: reauthenticate_api_v1_auth_reauthenticate_post
      parameters:
      - name: X-CSRF-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Csrf-Token
      - name: pm_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pm Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReauthenticateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Reauthenticate Api V1 Auth Reauthenticate Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/totp/setup:
    post:
      tags:
      - authentication
      summary: Setup Totp
      operationId: setup_totp_api_v1_auth_totp_setup_post
      parameters:
      - name: X-CSRF-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Csrf-Token
      - name: pm_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pm Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Setup Totp Api V1 Auth Totp Setup Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/totp/verify:
    post:
      tags:
      - authentication
      summary: Verify Totp Setup
      operationId: verify_totp_setup_api_v1_auth_totp_verify_post
      parameters:
      - name: code
        in: query
        required: true
        schema:
          type: string
          title: Code
      - name: X-CSRF-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Csrf-Token
      - name: pm_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pm Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: boolean
                title: Response Verify Totp Setup Api V1 Auth Totp Verify Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BootstrapRequest:
      properties:
        bootstrap_secret:
          type: string
          minLength: 16
          title: Bootstrap Secret
        email:
          type: string
          format: email
          title: Email
        display_name:
          type: string
          maxLength: 120
          minLength: 1
          title: Display Name
        password:
          type: string
          maxLength: 1024
          minLength: 14
          title: Password
      additionalProperties: false
      type: object
      required:
      - bootstrap_secret
      - email
      - display_name
      - password
      title: BootstrapRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          maxLength: 1024
          title: Password
        totp_code:
          anyOf:
          - type: string
            pattern: ^\d{6}$
          - type: 'null'
          title: Totp Code
      additionalProperties: false
      type: object
      required:
      - email
      - password
      title: LoginRequest
...
    ReauthenticateRequest:
      properties:
        password:
          type: string
          maxLength: 1024
          title: Password
        totp_code:
          anyOf:
          - type: string
            pattern: ^\d{6}$
          - type: 'null'
          title: Totp Code
      additionalProperties: false
      type: object
      required:
      - password
      title: ReauthenticateRequest
...
    SessionView:
      properties:
        authenticated:
          type: boolean
          title: Authenticated
        user:
          anyOf:
          - $ref: '#/components/schemas/UserSummary'
          - type: 'null'
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        csrf_token:
          anyOf:
          - type: string
          - type: 'null'
          title: Csrf Token
        bootstrap_required:
          type: boolean
          title: Bootstrap Required
          default: false
      additionalProperties: false
      type: object
      required:
      - authenticated
      title: SessionView
...
    UserSummary:
      properties:
        id:
          type: string
          title: Id
        email:
          type: string
          format: email
          title: Email
        display_name:
          type: string
          title: Display Name
        roles:
          items:
            type: string
          type: array
          title: Roles
        permissions:
          items:
            type: string
          type: array
          title: Permissions
        all_sites:
          type: boolean
          title: All Sites
          default: true
        site_ids:
          items:
            type: string
          type: array
          title: Site Ids
      additionalProperties: false
      type: object
      required:
      - id
      - email
      - display_name
      - roles
      title: UserSummary
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError